requirements.txt
# Core dependencies
docker~=7.1.0
mysql-connector-python>=8.3.0
cryptography>=42.0.5
python-dotenv>=1.0.1
astunparse>=1.6.3

# Security/additional packages
types-python-dateutil>=2.8.19
python-json-logger>=2.0.7

# Additional packages
requests>=2.31.0
pyyaml>=6.0.1
mysql-connector-python>=8.2.0
docker>=6.1.3
pytest>=7.4.3
